Exercises up to 4.21

+++ controllers/blogs.js
@@ -1,19 +1,31 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
     .populate('user', {username: 1, name: 1})
-    
+
   response.json(blogs)
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log(user)
+
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blogToRemove = await Blog.findById(request.params.id)
+    console.log(blogToRemove)
+    if ( blogToRemove.user.toString() === user._id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
   } catch (error) {
     next(error)
   }
@@ -47,7 +59,11 @@ blogsRouter.post('/', async (request, response, next) => {
     next(error)
   }
 
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -66,5 +82,7 @@ blogsRouter.post('/', async (request, response, next) => {
     next(exception)
   }
 })
+
+
 
 module.exports = blogsRouter
